requirements.txt
Flask==2.2.2
prometheus-client==0.16.0
prometheus-flask-exporter==0.21.0
python-dotenv==1.0.0
